server/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
colorama==0.4.6
fastapi==0.110.1
flake8==7.0.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.0
pluggy==1.5.0
pycodestyle==2.11.1
pydantic==2.7.0
pydantic_core==2.18.1
pyflakes==3.2.0
PyMySQL==1.1.1
pytest==8.2.1
python-dotenv==1.0.1
setuptools==70.0.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
wheel==0.43.0